pyproject.toml
[tool.poetry]
name = "streamlit-howest"
version = "0.1.0"
description = ""
authors = ["huyghebaertthomas <thomas.huyghebaert@howest.be>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
jupyter = "^1.1.1"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
streamlit-extras = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"